Fix(Field Processing): Commonize field processing components
