+++ alert-common/src/main/java/com/synopsys/integration/alert/common/descriptor/config/field/PasswordConfigField.java
@@ -22,7 +22,7 @@
  */
 package com.synopsys.integration.alert.common.descriptor.config.field;
 
-import com.synopsys.integration.alert.common.descriptor.config.field.validators.EncryptionValidator;
+import com.synopsys.integration.alert.common.descriptor.config.field.validation.EncryptionValidator;
 import com.synopsys.integration.alert.common.enumeration.FieldType;
 
 public class PasswordConfigField extends ConfigField {
