+++ alert-common/src/main/java/com/synopsys/integration/alert/common/descriptor/config/field/URLInputConfigField.java
@@ -27,7 +27,7 @@ import java.net.URL;
 
 import org.apache.commons.lang3.StringUtils;
 
-import com.synopsys.integration.alert.common.descriptor.config.field.validators.ValidationResult;
+import com.synopsys.integration.alert.common.descriptor.config.field.validation.ValidationResult;
 import com.synopsys.integration.alert.common.rest.model.FieldModel;
 import com.synopsys.integration.alert.common.rest.model.FieldValueModel;
