+++ alert-common/src/main/java/com/synopsys/integration/alert/common/descriptor/config/field/validation/UploadValidationFunction.java
@@ -20,7 +20,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.common.descriptor.config.field.validators;
+package com.synopsys.integration.alert.common.descriptor.config.field.validation;
 
-public abstract class EncryptionValidator implements ConfigValidationFunction {
+import java.io.File;
+import java.util.function.Function;
+
+@FunctionalInterface
+public interface UploadValidationFunction extends Function<File, ValidationResult> {
 }
