+++ alert-common/src/main/java/com/synopsys/integration/alert/common/descriptor/config/field/validation/ValidationResult.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.common.descriptor.config.field.validators;
+package com.synopsys.integration.alert.common.descriptor.config.field.validation;
 
 import java.util.Arrays;
 import java.util.Collection;
