+++ src/main/java/com/synopsys/integration/alert/channel/azure/boards/oauth/AzureOAuthAuthenticateValidator.java
@@ -25,14 +25,14 @@ package com.synopsys.integration.alert.channel.azure.boards.oauth;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.synopsys.integration.alert.common.descriptor.config.field.validators.ConfigValidationFunction;
-import com.synopsys.integration.alert.common.descriptor.config.field.validators.ValidationResult;
+import com.synopsys.integration.alert.common.descriptor.config.field.validation.ConfigValidationFunction;
+import com.synopsys.integration.alert.common.descriptor.config.field.validation.ValidationResult;
 import com.synopsys.integration.alert.common.rest.model.FieldModel;
 import com.synopsys.integration.alert.common.rest.model.FieldValueModel;
 
 @Component
 public class AzureOAuthAuthenticateValidator implements ConfigValidationFunction {
-    private OAuthRequestValidator oAuthRequestValidator;
+    private final OAuthRequestValidator oAuthRequestValidator;
 
     @Autowired
     public AzureOAuthAuthenticateValidator(OAuthRequestValidator oAuthRequestValidator) {
