+++ src/main/java/com/synopsys/integration/alert/channel/jira/server/web/JiraServerCustomFunctionAction.java
@@ -61,7 +61,7 @@ public class JiraServerCustomFunctionAction extends CustomFunctionAction<String>
 
     @Autowired
     public JiraServerCustomFunctionAction(AuthorizationManager authorizationManager, JiraServerChannelKey jiraChannelKey, ConfigurationAccessor configurationAccessor, Gson gson) {
-        super(authorizationManager);
+        super(authorizationManager, fieldModelProcessor);
         this.jiraChannelKey = jiraChannelKey;
         this.configurationAccessor = configurationAccessor;
         this.gson = gson;
