+++ src/main/java/com/synopsys/integration/alert/provider/blackduck/web/VulnerabilityNotificationFilterCustomFunctionAction.java
@@ -51,7 +51,7 @@ public class VulnerabilityNotificationFilterCustomFunctionAction extends CustomF
 
     @Autowired
     public VulnerabilityNotificationFilterCustomFunctionAction(AuthorizationManager authorizationManager) {
-        super(authorizationManager);
+        super(authorizationManager, fieldModelProcessor);
     }
 
     @Override
